move tests back to client because they're mocking a client interface
